==> 00-B/18-Improving Query Performance in SQL Server/01-Introduction, Review and The Order of Things.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import sqlite3


engine = sqlite3.connect('example.db')
def rsq(query: str) -> pd.DataFrame:
    return pd.read_sql_query(query, engine)

SELECT 
    CASE
        WHEN FROM_UNIXTIME(UNIX_TIMESTAMP('121212', 'yyMMdd')) > DATE_ADD(CURRENT_DATE,-30*30)
            THEN 'greater than'
        ELSE 'less than or equal to'
    END as compare_result;

rsq("""SELECT 
    CASE
        WHEN strftime('%s', '121212') > strftime('%s', 'now', '-30 days')
            THEN 'greater than'
        ELSE 'less than or equal to'
    END as compare_result;""")



# SELECT *
# FROM Nations
# ORDER BY CountryCode
# OFFSET 100 ROWS
# FETCH NEXT 100 ROWS ONLY; 

df = pd.read_csv('./earthquake_1995-2023.csv')
df1 = pd.read_csv('./worldcities.csv')
# df = df[['iso2', 'city_ascii','population','lat', 'lng']]
# df.rename(columns={'iso2': 'CountryCode', 'city_ascii': 'CityName', 'population': 'Pop2017', 'lat': 'Latitude', 'lng': 'Longitude'}, inplace=True)
df['Date'] = pd.to_datetime(df['date_time'],dayfirst=True).dt.date
df['Time'] = pd.to_datetime(df['date_time'],dayfirst=True).dt.strftime('%H:%M:%S')
df.rename(columns={'date_time': 'DateTime', 'longitude': 'Longitude', 'latitude': 'Latitude', 'depth': 'Depth', 'magnitude': 'Magnitude', 'title': 'Title', 'country': 'Country'}, inplace=True)
df = df[['Date', 'Time', 'Latitude', 'Longitude', 'Depth', 'Magnitude', 'Title', 'Country','location']]
# [Date	Time	latitude	longitude	depth	magnitude	place	NearestPop	Country]]
   # ['title', 'magnitude', 'date_time', 'cdi', 'mmi', 'alert', 'tsunami',
    #    'sig', 'net', 'nst', 'dmin', 'gap', 'magType', 'depth', 'latitude',
    #    'longitude', 'location', 'continent', 'country'],
df['country'] = df['Country'].fillna(df['location'].str.split(',').str[-1].str.strip())
df['NearestPop'] = df['location'].str.split(',').str[0]
# df['Country'] = 
df = pd.merge(df,df1,right_on='city_ascii',left_on='NearestPop',how='left')[['Date', 'Time', 'Latitude', 'Longitude', 'Depth', 'Magnitude', 'location', 'NearestPop','iso2']]
df.rename(columns={'iso2': 'Country','location':'place'}, inplace=True)
df.drop_duplicates(subset=['place','Date','Time'],inplace=True)
df   
df = pd.read_clipboard()
df.to_sql('Orders', engine, if_exists='append', index=False)

select PlayerName, Country,
round(Weight_kg/SQUARE(Height_cm/100),2) BMI
from Players Where Country = 'USA'
Or Country = 'Canada'
order by BMI;

SELECT PlayerName, Country,
ROUND(Weight_kg/SQUARE(Height_cm/100),2) BMI
FROM Players
WHERE Country = 'USA'
OR Country = 'Canada'
ORDER BY BMI;

rsq("""select PlayerName, Country,
round(Weight_kg/POWER(Height_cm/100, 2), 2) BMI
from Players Where Country = 'USA'
Or Country = 'Canada'
order by BMI;""")

rsq("""-- Your friend's query
-- First attempt, contains errors and inconsistent formatting
/*
select PlayerName, p.Country,sum(ps.TotalPoints)
AS TotalPoints
FROM PlayerStats ps inner join Players On ps.PlayerName = p.PlayerName
WHERE p.Country = 'New Zeeland'
Group
by PlayerName, Country
order by Country;
*/

-- Your query
-- Second attempt - errors corrected and formatting fixed

SELECT p.PlayerName, p.Country,
SUM(ps.TotalPoints) AS TotalPoints
FROM PlayerStats ps
INNER JOIN Players p
ON ps.PlayerName = p.PlayerName
WHERE p.Country = 'New Zealand'
GROUP BY p.PlayerName, p.Country;""")

rsq("""SELECT p.PlayerName, p.Country,
SUM(ps.TotalPoints) AS TotalPoints
FROM PlayerStats ps
INNER JOIN Players p
ON ps.PlayerName = p.PlayerName
WHERE p.Country = 'Australia'
GROUP BY p.PlayerName, p.Country""")

SELECT Team,
ROUND(AVG(BMI),2) AS AvgTeamBMI -- Alias the new column
FROM PlayerStats AS ps -- Alias PlayerStats table
INNER JOIN
(SELECT PlayerName, Country,
Weight_kg/SQUARE(Height_cm/100) BMI
FROM Players) AS p -- Alias the sub-query
-- Alias the joining columns
ON p.PlayerName = ps.PlayerName
GROUP BY Team
HAVING AVG(BMI) >= 25;

rsq("""SELECT Team,
ROUND(AVG(BMI),2) AS AvgTeamBMI -- Alias the new column
FROM PlayerStats AS ps -- Alias PlayerStats table
INNER JOIN
(SELECT PlayerName, Country,
Weight_kg/POWER(Height_cm/100, 2) BMI
FROM Players) AS p -- Alias the sub-query
-- Alias the joining columns
ON p.PlayerName = ps.PlayerName
GROUP BY Team
HAVING AVG(BMI) >= 25;""")

rsq("""SELECT date, place, Depth, magnitude
FROM Earthquakes
WHERE magnitude > 6
    AND Depth > 100
ORDER BY Depth DESC;""")


rsq("""SELECT -- Limit the number of rows to ten
*
FROM Earthquakes
WHERE country = 'PG'
OR Country = 'ID'
ORDER BY Depth; -- Order results from shallowest to deepest""")

rsq("""SELECT Date, Country, Place, Depth, Magnitude
FROM Earthquakes
WHERE Magnitude > 8
AND Depth > 500
ORDER BY Depth DESC;""")

rsq("""/*
Returns earthquakes in New Zealand with a magnitude of 7.5 or more
*/

SELECT Date, Place, NearestPop, Magnitude
FROM Earthquakes
WHERE Country = 'NZ'
AND Magnitude >= 7.5
ORDER BY Magnitude DESC;""")

rsq("""SELECT PlayerName,
Team,
Position,
(OffRebound+DefRebound)/CAST(GamesPlayed AS numeric) AS AvgRebounds
FROM PlayerStats
WHERE AvgRebounds >= 12;""")

rsq("""SELECT PlayerName,
Country,
College,
DraftYear,
DraftNumber
FROM Players
-- WHERE UPPER(LEFT(College,5)) LIKE 'LOU%'
WHERE College LIKE 'Louisiana%'; -- Add the wildcard filter""")

rsq("""SELECT Country, COUNT(*) CountOfPlayers
FROM Players
GROUP BY Country
HAVING Country
IN ('Argentina','Brazil','Dominican Republic'
,'Puerto Rico');""")
