==> 00-B/18-Improving Query Performance in SQL Server/02-Filtering and Data Interrogation.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import sqlite3


engine = sqlite3.connect('example.db')
def rsq(query: str) -> pd.DataFrame:
    return pd.read_sql_query(query, engine)

rsq("""-- Second query

-- Add the new column to the select statement
SELECT PlayerName, 
       Team, 
       Position, 
       AvgRebounds -- Add the new column
FROM
     -- Sub-query starts here                             
	(SELECT 
      PlayerName, 
      Team, 
      Position,
      -- Calculate average total rebounds
     (OffRebound+DefRebound)/CAST(GamesPlayed AS numeric) AS AvgRebounds
	 FROM PlayerStats) as tr
WHERE AvgRebounds >= 12; -- Filter rows""")

rsq("""SELECT PlayerName,
Country,
College,
DraftYear,
DraftNumber
FROM Players
-- WHERE UPPER(LEFT(College,5)) LIKE 'LOU%'
WHERE College LIKE 'Louisiana%'; -- Add the wildcard filter""")

rsq("""SELECT Country, COUNT(*) CountOfPlayers
FROM Players
GROUP BY Country
HAVING Country --- not good practice
IN ('Argentina','Brazil','Dominican Republic'
,'Puerto Rico');""")

rsq("""SELECT Country, COUNT(*) CountOfPlayers
FROM Players
-- Add the filter condition
WHERE Country
-- Fill in the missing countries
IN ('Argentina','Brazil','Dominican Republic'
,'Puerto Rico')
GROUP BY Country;""")

rsq("""SELECT Team,
SUM(TotalPoints) AS TotalPFPoints
FROM PlayerStats
-- Filter for only rows with power forwards
WHERE Position = 'PF'
GROUP BY Team
-- Filter for total points greater than 3000
HAVING SUM(TotalPoints) > 3000;""")

rsq("""SELECT Team,
SUM(TotalPoints) AS TotalCPoints
FROM PlayerStats
WHERE Position = 'C'
GROUP BY Team
HAVING SUM(TotalPoints) > 2500;""")

SELECT TOP 25 PERCENT -- Limit rows to the upper quartile
Latitude,
Longitude,
Magnitude,
Depth,
NearestPop
FROM Earthquakes
WHERE Country = 'PG'
OR Country = 'ID'
ORDER BY Magnitude DESC; -- Order the results

rsq("""SELECT Latitude,
Latitude,
Longitude,
Magnitude,
Depth,
NearestPop
FROM Earthquakes
WHERE Country = 'PG'
OR Country = 'ID'
ORDER BY Magnitude DESC -- Order the results
LIMIT (SELECT COUNT(*) FROM Earthquakes) / 4;""")


rsq("""SELECT NearestPop,
Country,
COUNT(NearestPop) NumEarthquakes -- Number of cities
FROM Earthquakes
WHERE Magnitude >= 8
AND Country IS NOT NULL
GROUP BY NearestPop, Country -- Group columns
ORDER BY NumEarthquakes DESC;""")
