==> 00-B/18-Improving Query Performance in SQL Server/03-Sub-queries and presence or absence.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import sqlite3


engine = sqlite3.connect('example.db')
def rsq(query: str) -> pd.DataFrame:
    return pd.read_sql_query(query, engine)

rsq("""SELECT UNContinentRegion,
CountryName,
(SELECT AVG(Magnitude) -- Add average magnitude
FROM Earthquakes e
-- Add country code reference
WHERE n.Code2 = e.Country) AS AverageMagnitude
FROM Nations n
ORDER BY UNContinentRegion DESC,
AverageMagnitude DESC;""")

rsq("""SELECT WorldBankRegion,
CountryName
FROM Nations
WHERE Code2 IN -- Add the operator to compare queries
(SELECT CountryCode -- Country code column
FROM Cities);""")

rsq("""SELECT WorldBankRegion,
CountryName,
Code2,
Capital, -- Country capital column
Pop2017
FROM Nations AS n
WHERE NOT EXISTS -- Add the operator to compare queries
(SELECT 1
FROM Cities AS c
WHERE n.Code2 = c.CountryCode); -- Columns being compared""")

rsq("""SELECT WorldBankRegion,
CountryName
FROM Nations
WHERE Code2 NOT IN -- Add the operator to compare queries
(SELECT CountryCode -- Country code column
FROM Cities);""")

rsq("""SELECT WorldBankRegion,
CountryName,
Capital
FROM Nations
WHERE Capital NOT IN
(SELECT NearestPop
FROM Earthquakes
WHERE NearestPop IS NOT NULL); -- filter condition""")

rsq("""SELECT WorldBankRegion,
CountryName,
Capital -- Capital city name column
FROM Nations
WHERE Capital  IN
(SELECT NearestPop -- City name column
FROM Earthquakes);""")

rsq("""-- Initial query
SELECT TeamName,
TeamCode,
City
FROM Teams AS t -- Add table
WHERE EXISTS -- Operator to compare queries
(SELECT 1
FROM Earthquakes AS e -- Add table
WHERE t.City = e.NearestPop);""")

rsq("""-- Second query
SELECT t.TeamName,
t.TeamCode,
t.City,
e.Date,
e.Place, -- Place description
e.Country -- Country code
FROM Teams AS t
INNER JOIN Earthquakes AS e -- Operator to compare tables
ON t.City = e.NearestPop""")

# SELECT * 
# FROM Orders
# order by OrderID
# offset 300 rows

rsq("""-- First attempt
SELECT c.CustomerID,
c.CompanyName,
c.ContactName,
c.ContactTitle,
c.Phone
FROM Customers c
LEFT OUTER JOIN Orders o -- Joining operator
ON c.CustomerID = o.CustomerID -- Joining columns
WHERE c.Country = 'France';""")

rsq("""-- Second attempt
SELECT c.CustomerID,
c.CompanyName,
c.ContactName,
c.ContactTitle,
c.Phone
FROM Customers c
LEFT OUTER JOIN Orders o
ON c.CustomerID = o.CustomerID
WHERE c.Country = 'France'
AND o.CustomerID IS NULL; -- Filter condition
""")

rsq("""-- Second attempt
SELECT c.CustomerID,
c.CompanyName,
c.ContactName,
c.ContactTitle,
c.Phone
FROM Customers c
LEFT OUTER JOIN Orders o
ON c.CustomerID = o.CustomerID
WHERE c.Country = 'France'
AND o.CustomerID IS NULL; -- Filter condition

Test your knowle""")
