==> 00-B/18-Improving Query Performance in SQL Server/04-Query performance tuning.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import sqlite3


engine = sqlite3.connect('example.db')
def rsq(query: str) -> pd.DataFrame:
    return pd.read_sql_query(query, engine)

SET STATISTICS TIME ON

rsq("""EXPLAIN QUERY PLAN SELECT *
FROM Teams
-- Sub-query 1
WHERE City IN -- Sub-query filter operator
(SELECT CityName
FROM Cities) -- Table from Earthquakes database
-- Sub-query 2
AND City IN -- Sub-query filter operator
(SELECT CityName
FROM Cities
WHERE CountryCode IN ('US','CA'))
-- Sub-query 3
AND City IN -- Sub-query filter operator
(SELECT CityName
FROM Cities
WHERE Pop2017 >2000000);""")


rsq("""-- Query 1
SELECT *
FROM Teams
-- Sub-query 1
WHERE City IN -- Sub-query filter operator
(SELECT CityName
FROM Cities) -- Table from Earthquakes database
-- Sub-query 2
AND City IN -- Sub-query filter operator
(SELECT CityName
FROM Cities
WHERE CountryCode IN ('US','CA'))
-- Sub-query 3
AND City IN -- Sub-query filter operator
(SELECT CityName
FROM Cities
WHERE Pop2017 >2000000);""")

rsq("""-- Query 2
SELECT *
FROM Teams AS t
WHERE EXISTS -- Sub-query filter operator
(SELECT 1
FROM Cities AS c
WHERE t.City = c.CityName -- Columns being compared
AND c.CountryCode IN ('US','CA')
AND c.Pop2017 > 2000000);""")

SET STATISTICS TIME OFF -- Turn the time command off

SET STATISTICS IO ON -- Turn the IO command on

rsq("""-- Example 1
SELECT CustomerID,
CompanyName,
(SELECT COUNT(*)
FROM Orders AS o -- Add table
WHERE c.CustomerID = o.CustomerID) CountOrders
FROM Customers AS c
WHERE CustomerID IN -- Add filter operator
(SELECT CustomerID
FROM Orders
WHERE ShipCity IN
('Berlin','Bern','Bruxelles','Helsinki',
'Lisboa','Madrid','Paris','London'));""")

rsq("""-- Example 2
SELECT c.CustomerID,
c.CompanyName,
COUNT(o.CustomerID)
FROM Customers AS c
INNER JOIN Orders AS o -- Join operator
ON c.CustomerID = o.CustomerID
WHERE o.ShipCity IN -- Shipping destination column
('Berlin','Bern','Bruxelles','Helsinki',
'Lisboa','Madrid','Paris','London')
GROUP BY c.CustomerID,
c.CompanyName;""")

SET STATISTICS IO OFF -- Turn the IO command off

rsq("""-- Query 1
SELECT *
FROM Cities
WHERE CountryCode = 'RU' -- Country code
OR CountryCode = 'CN' -- Country code""")

rsq("""-- Query 2
SELECT *
FROM Cities
WHERE CountryCode IN ('JM','NZ') -- Country codes""")

rsq("""SELECT CityName AS NearCityName,
CountryCode
FROM Cities

UNION -- Append queries

SELECT Capital AS NearCityName,
Code2 AS CountryCode
FROM Nations;""")
